int to long

+++ src/SalesManager.java
@@ -1,13 +1,13 @@
 public class SalesManager {
-    protected int[] sales; // массив продаж
+    protected long[] sales; // массив продаж
 
-    public SalesManager(int[] sales) { //конструктор менеджера по продажам
+    public SalesManager(long[] sales) { //конструктор менеджера по продажам
         this.sales = sales;
     }
 
-    public int max() {
-        int max = -1;
-        for (int sale : sales) {
+    public long max() {
+        long max = -1;
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
